Created ErrorBoundary folder inside ui/components/shared/ & moved ErrorBoundary.tsx to ui/components/shared/ErrorBoundary/
